Attempt to automate the release
On a merge to `main` and a successful run of the CI we attempt to build
the code and commit it.  We merge into `release` the `main` branch so
that it's reasonably easy to see the history.  The additional merge
commits may look a bit ugly but it's not the end of the world for this
branch.

We have to use `--allow-unrelated-histories` since we have a shallow
checkout and git doesn't know that these branches are related.