Update README and add more useful inputs
Update the README with examples, documentation, and an FAQ.  During
writing this, I discovered a few omissions in configuration options.

`build-configuration` specifies an additional cache string that can
differentiate the ninja cache between different builds on the same
platform.

`affected` can be used to augment the list of changed files from `git
diff` so that `trimja` will always keep a particular set of build
outputs.